+++ src/templates/Blog.jsx
@@ -1,13 +1,13 @@
 import React from "react";
 import { graphql } from "gatsby";
 import { GatsbyImage } from "gatsby-plugin-image";
+import moment from "moment";
 import Layout from "../layout";
-import PostTags from "../components/PostTags/PostTags";
 import SocialLinks from "../components/SocialLinks/SocialLinks";
 import SEO from "../components/SEO/SEO";
 import config from "../../data/SiteConfig";
 
-export default class ModuleTemplate extends React.Component {
+export default class Blog extends React.Component {
   render() {
     const { data, pageContext } = this.props;
     const { slug } = pageContext;
@@ -19,24 +19,32 @@ export default class ModuleTemplate extends React.Component {
     if (!post.category_id) {
       post.category_id = config.postDefaultCategoryID;
     }
+    const { date } = postNode.fields;
     return (
       <Layout>
         <div>
           <div>
+            <div className="title">
+              <h1>{post.title}</h1>
+              <h6>
+                {`Posted by ${post.author} on `}
+                <span>{moment(date).format("MMMM DD, YYYY")}</span>
+              </h6>
+            </div>
+            <br />
             <GatsbyImage
               className="post-cover"
               image={post.cover.childImageSharp.gatsbyImageData}
             />
-            <h1>{post.title}</h1>
-            <div className="d-flex mt-2 ml-2">
-              {post.tags.map((tag) => (
-                <PostTags tags={tag} type="modules" />
-              ))}
-            </div>
+            <br />
             <hr />
-            <div dangerouslySetInnerHTML={{ __html: postNode.html }} />
+            <div
+              className="post-content"
+              dangerouslySetInnerHTML={{ __html: postNode.html }}
+            />
+            <hr />
             <div className="post-meta">
-              <SocialLinks postPath={`/modules${slug}`} postNode={postNode} />
+              <SocialLinks postPath={`/blog${slug}`} postNode={postNode} />
             </div>
           </div>
         </div>
@@ -47,21 +55,21 @@ export default class ModuleTemplate extends React.Component {
 
 /* eslint no-undef: "off" */
 export const pageQuery = graphql`
-  query ModulesBySlug($slug: String!) {
+  query BlogPostBySlug($slug: String!) {
     markdownRemark(fields: { slug: { eq: $slug } }) {
       html
       timeToRead
       excerpt
       frontmatter {
         title
+        tags
+        author
         cover {
           publicURL
           childImageSharp {
             gatsbyImageData
           }
         }
-        date
-        tags
       }
       fields {
         slug
